refactor game main page
